refactor(hdi): make the v2 validation API canonical; delete legacy op/flat_op
Flip hdi's prelude to export op_v2/flat_op_v2 and the v2
holochain_integrity_types::dht_v2::{Op, Action, Record} in place of the
legacy per-variant equivalents, with explicit re-exports to shadow the
still-present legacy glob (Op, its 7 variant structs, Action, Record,
ActionType). Delete the now-unused legacy op.rs/flat_op.rs modules and
their tests.

Move op_v2's remaining dependency on hdi's legacy op.rs (generic
entry/link-type resolution helpers, unrelated to the legacy Action
shape) into op_v2.rs itself. Fix hash_action to hash the v2 Action, and
trim the legacy-Action-shaped test short-hand builders and integration
test that have no v2 use.
